Add files via upload
The idea presented in the previous section can be extended to a bidirectional configuration. Fig. shows a 3-bit bidirectional capacitive DAC. In the bidirectional DAC, negative and the positive references are used, respectively. The 2 input nodes to the comparator are compared in each iteration, so if (VXP – VXN) > 0, then the comparator will output a 1, otherwise, it produces a 0. The bidirectional configuration has many advantages over the single-ended configuration like better common-mode rejection ratio (CMRR), the reduction of even harmonic distortion, and the doubling of input voltage range, but at the expense of larger layout area since most of the components are doubled for the bidirectional configuration.